Check points added
In T1: Changed date format and removed the time here.  and exporting as the xlsx format.

In T2: Which is used to check the total case count for the report usage.

In T3: We have no changes.

In T4: Here we are going to merge both the CSV and Excel file

In T5: This open the xlsx file and load to the local storage and delete the xlsx file and start converting to the tuple

In T6: Original from the thinakaran